refactor(analytics): extract telemetry useEffects into adjacent hook files
Moves the four heaviest telemetry useEffects out of the business
components and into co-located *.telemetry.ts / *.hooks.ts files:

- LensSearchDialog → useSearchTelemetry (debounced search event)
- LensFilters → useFiltersTelemetry (debounced filter_apply/reset)
- CompareTelemetry → useCompareViewTelemetry + useCompareScrollTelemetry
- LensDetailTelemetry → useLensView/Scroll/DwellTelemetry

Single-line track() calls in onClick/onSubmit handlers are left in
place — extracting those would add a hook indirection without
shortening the call site.
